store sync settings in vault
